test(app-e2e): split elk optimizeDeps regressions into their own file
